requirements.txt
langchain==1.1.3
langchain-openai==1.1.1
python-dotenv==1.2.1
langchain-deepseek==1.0.1
langgraph-cli[inmem]==0.4.8
tavily-python==0.7.15
sqlalchemy
pymysql
loguru

# Note: There's a dependency conflict between fireworks-ai (requires protobuf==5.29.3)
# and langgraph-cli[inmem] (requires protobuf>=6.32.1). Currently protobuf 6.33.2 is installed.
# If fireworks-ai functionality is needed, consider upgrading langgraph-cli or waiting
# for fireworks-ai to support protobuf 6.x